+++ api/v1alpha1/conversion_test.go
@@ -14,7 +14,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	runtimeserializer "k8s.io/apimachinery/pkg/runtime/serializer"
 
-	"github.com/vmware-tanzu/vm-operator/api/utilconversion"
+	"github.com/vmware-tanzu/vm-operator/api/utilconversion/fuzztests"
 	vmopv1a1 "github.com/vmware-tanzu/vm-operator/api/v1alpha1"
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha4"
 )
@@ -23,7 +23,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 
 	var (
 		scheme *runtime.Scheme
-		input  utilconversion.FuzzTestFuncInput
+		input  fuzztests.FuzzTestFuncInput
 	)
 
 	BeforeEach(func() {
@@ -33,12 +33,12 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 	})
 
 	AfterEach(func() {
-		input = utilconversion.FuzzTestFuncInput{}
+		input = fuzztests.FuzzTestFuncInput{}
 	})
 
 	Context("VirtualMachine", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachine{},
 				Spoke:  &vmopv1a1.VirtualMachine{},
@@ -49,19 +49,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineClass", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineClass{},
 				Spoke:  &vmopv1a1.VirtualMachineClass{},
@@ -72,19 +72,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineImage", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineImage{},
 				Spoke:  &vmopv1a1.VirtualMachineImage{},
@@ -95,19 +95,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("ClusterVirtualMachineImage", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.ClusterVirtualMachineImage{},
 				Spoke:  &vmopv1a1.ClusterVirtualMachineImage{},
@@ -118,19 +118,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachinePublishRequest", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachinePublishRequest{},
 				Spoke:  &vmopv1a1.VirtualMachinePublishRequest{},
@@ -141,19 +141,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineService", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineService{},
 				Spoke:  &vmopv1a1.VirtualMachineService{},
@@ -161,18 +161,18 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 	Context("VirtualMachineSetResourcePolicy", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineSetResourcePolicy{},
 				Spoke:  &vmopv1a1.VirtualMachineSetResourcePolicy{},
@@ -180,12 +180,12 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
