+++ api/v1alpha2/conversion_test.go
@@ -16,7 +16,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	runtimeserializer "k8s.io/apimachinery/pkg/runtime/serializer"
 
-	"github.com/vmware-tanzu/vm-operator/api/utilconversion"
+	"github.com/vmware-tanzu/vm-operator/api/utilconversion/fuzztests"
 	vmopv1a2 "github.com/vmware-tanzu/vm-operator/api/v1alpha2"
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha4"
 	vmopv1sysprep "github.com/vmware-tanzu/vm-operator/api/v1alpha4/sysprep"
@@ -26,7 +26,7 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 
 	var (
 		scheme *runtime.Scheme
-		input  utilconversion.FuzzTestFuncInput
+		input  fuzztests.FuzzTestFuncInput
 	)
 
 	BeforeEach(func() {
@@ -36,12 +36,12 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 	})
 
 	AfterEach(func() {
-		input = utilconversion.FuzzTestFuncInput{}
+		input = fuzztests.FuzzTestFuncInput{}
 	})
 
 	Context("VirtualMachine", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachine{},
 				Spoke:  &vmopv1a2.VirtualMachine{},
@@ -52,19 +52,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineClass", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineClass{},
 				Spoke:  &vmopv1a2.VirtualMachineClass{},
@@ -72,19 +72,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineImage", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineImage{},
 				Spoke:  &vmopv1a2.VirtualMachineImage{},
@@ -95,19 +95,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("ClusterVirtualMachineImage", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.ClusterVirtualMachineImage{},
 				Spoke:  &vmopv1a2.ClusterVirtualMachineImage{},
@@ -118,19 +118,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachinePublishRequest", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachinePublishRequest{},
 				Spoke:  &vmopv1a2.VirtualMachinePublishRequest{},
@@ -138,19 +138,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineService", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineService{},
 				Spoke:  &vmopv1a2.VirtualMachineService{},
@@ -158,19 +158,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineSetResourcePolicy", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineSetResourcePolicy{},
 				Spoke:  &vmopv1a2.VirtualMachineSetResourcePolicy{},
@@ -178,19 +178,19 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
 
 	Context("VirtualMachineWebConsoleRequest", func() {
 		BeforeEach(func() {
-			input = utilconversion.FuzzTestFuncInput{
+			input = fuzztests.FuzzTestFuncInput{
 				Scheme: scheme,
 				Hub:    &vmopv1.VirtualMachineWebConsoleRequest{},
 				Spoke:  &vmopv1a2.VirtualMachineWebConsoleRequest{},
@@ -198,12 +198,12 @@ var _ = Describe("FuzzyConversion", Label("api", "fuzz"), func() {
 		})
 		Context("Spoke-Hub-Spoke", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.SpokeHubSpoke(input)
+				fuzztests.SpokeHubSpoke(input)
 			})
 		})
 		Context("Hub-Spoke-Hub", func() {
 			It("should get fuzzy with it", func() {
-				utilconversion.HubSpokeHub(input)
+				fuzztests.HubSpokeHub(input)
 			})
 		})
 	})
